Move main logic to singletons
