+++ js/src/traceable.ts
@@ -9,7 +9,11 @@ import {
 } from "./run_trees.js";
 import { InvocationParamsSchema, KVMap } from "./schemas.js";
 import { isTracingEnabled } from "./env.js";
-import { TraceableLocalStorageContext } from "./singletons/traceable.js";
+import {
+  ROOT,
+  TraceableFunction,
+  TraceableLocalStorageContext,
+} from "./singletons/traceable.js";
 
 function isPromiseMethod(
   x: string | symbol
@@ -23,83 +27,6 @@ function isPromiseMethod(
 const asyncLocalStorage = TraceableLocalStorageContext.register(
   new AsyncLocalStorage<RunTree | undefined>()
 );
-
-export const ROOT = Symbol("langsmith:traceable:root");
-
-export type RunTreeLike = RunTree;
-
-type SmartPromise<T> = T extends AsyncGenerator
-  ? T
-  : T extends Promise<unknown>
-  ? T
-  : Promise<T>;
-
-type WrapArgReturnPair<Pair> = Pair extends [
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  infer Args extends any[],
-  infer Return
-]
-  ? Args extends [RunTreeLike, ...infer RestArgs]
-    ? {
-        (
-          runTree: RunTreeLike | typeof ROOT,
-          ...args: RestArgs
-        ): SmartPromise<Return>;
-        (config: RunnableConfigLike, ...args: RestArgs): SmartPromise<Return>;
-      }
-    : {
-        (...args: Args): SmartPromise<Return>;
-        (runTree: RunTreeLike, ...rest: Args): SmartPromise<Return>;
-        (config: RunnableConfigLike, ...args: Args): SmartPromise<Return>;
-      }
-  : never;
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-type UnionToIntersection<U> = (U extends any ? (x: U) => void : never) extends (
-  x: infer I
-) => void
-  ? I
-  : never;
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export type TraceableFunction<Func extends (...args: any[]) => any> =
-  // function overloads are represented as intersections rather than unions
-  // matches the behavior introduced in https://github.com/microsoft/TypeScript/pull/54448
-  Func extends {
-    (...args: infer A1): infer R1;
-    (...args: infer A2): infer R2;
-    (...args: infer A3): infer R3;
-    (...args: infer A4): infer R4;
-    (...args: infer A5): infer R5;
-  }
-    ? UnionToIntersection<
-        WrapArgReturnPair<[A1, R1] | [A2, R2] | [A3, R3] | [A4, R4] | [A5, R5]>
-      >
-    : Func extends {
-        (...args: infer A1): infer R1;
-        (...args: infer A2): infer R2;
-        (...args: infer A3): infer R3;
-        (...args: infer A4): infer R4;
-      }
-    ? UnionToIntersection<
-        WrapArgReturnPair<[A1, R1] | [A2, R2] | [A3, R3] | [A4, R4]>
-      >
-    : Func extends {
-        (...args: infer A1): infer R1;
-        (...args: infer A2): infer R2;
-        (...args: infer A3): infer R3;
-      }
-    ? UnionToIntersection<WrapArgReturnPair<[A1, R1] | [A2, R2] | [A3, R3]>>
-    : Func extends {
-        (...args: infer A1): infer R1;
-        (...args: infer A2): infer R2;
-      }
-    ? UnionToIntersection<WrapArgReturnPair<[A1, R1] | [A2, R2]>>
-    : Func extends {
-        (...args: infer A1): infer R1;
-      }
-    ? UnionToIntersection<WrapArgReturnPair<[A1, R1]>>
-    : never;
 
 const isAsyncIterable = (x: unknown): x is AsyncIterable<unknown> =>
   x != null &&
@@ -399,7 +326,7 @@ export function traceable<Func extends (...args: any[]) => any>(
   const { aggregator, argsConfigPath, ...runTreeConfig } = config ?? {};
 
   const traceableFunc = (
-    ...args: Inputs | [RunTreeLike, ...Inputs] | [RunnableConfigLike, ...Inputs]
+    ...args: Inputs | [RunTree, ...Inputs] | [RunnableConfigLike, ...Inputs]
   ) => {
     let ensuredConfig: RunTreeConfig;
     try {
@@ -680,14 +607,13 @@ export function traceable<Func extends (...args: any[]) => any>(
   return traceableFunc as TraceableFunction<Func>;
 }
 
-export { getCurrentRunTree } from "./singletons/traceable.js";
-
-export function isTraceableFunction(
-  x: unknown
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-): x is TraceableFunction<any> {
-  return typeof x === "function" && "langsmith:traceable" in x;
-}
+export {
+  type TraceableFunction,
+  getCurrentRunTree,
+  isTraceableFunction,
+  ROOT,
+} from "./singletons/traceable.js";
+export type RunTreeLike = RunTree;
 
 function isKVMap(x: unknown): x is Record<string, unknown> {
   if (typeof x !== "object" || x == null) {
